SQL Injection Fix
You may want to look at implementing this ASAP.

We had a couple of inividuals injection code into our database by using the string arugments passed from the client to the server.

This change essentially turns the string arguments recieved from the html document to number and boolean type and then checks if they're still number/boolean on the server once the data has been passed from the client to the server.